Steganography_Project/Steganography_Project/requirements.txt
Flask==3.0.0
pymongo==4.7.0
flask-jwt-extended==4.6.0
flask-cors==4.0.0
Pillow==10.3.0
Werkzeug==3.0.1
